requirements.txt
anthropic>=0.84.0
pydantic>=2.0
python-dotenv>=1.0
rich>=13.0
pymupdf>=1.24.0
arxiv>=2.1.0
requests>=2.28.0